Move old interp/fourier to new folder structure
