+++ src/python/zquantum/qaoa/parameter_initialization/__init__.py
@@ -1,1 +1,3 @@
 from ._fourier import FourierOptimizer
+from ._interp import get_new_layer_params_using_interp
+from ._old_fourier import Fourier
